# react/public/openapi3_1.yaml
openapi: "3.1.0"
info:
  version: "1.0.0"
  title: "Event ticketing app"
  description: "Project of a web application for selling event tickets using blockchain."
servers:
  - url: http://localhost:{port}/api
    description: Development server
    variables:
      port:
        enum:
          - "3000"
        default: "3000"
  - url: https://jalowiecki.toadres.pl/api
    description: Production server
paths:
...
  /artists/{id}:
    parameters:
      - name: id
        description: The unique identifier of the artist
        in: path
        schema:
          $ref: "#/components/schemas/Id"
        required: true
    get:
      operationId: GetArtist
      description: Get artist details
      responses:
        "200":
          description: Artist details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetArtist"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Artists
    patch:
      operationId: UpdateArtist
      description: Update artist details
      security:
        - bearerAuth: [ADMINISTRATOR]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtist"
      responses:
        "200":
          description: Artist updated details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetArtist"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Artists
...
  /uploads:
    post:
      operationId: CreateUpload
      description: Create new upload
      security:
        - bearerAuth: [EVENTS_ORGANIZER, ADMINISTRATOR]
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateUpload"
      responses:
        "201":
          description: Created upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUpload"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Uploads
...
  /reviews/to-approve:
    get:
      operationId: ListReviewsToApprove
      description: List reviews to approve
      security:
        - bearerAuth: [ADMINISTRATOR]
      parameters:
        - name: page
          description: A listing page number
          in: query
          schema:
            $ref: "#/components/schemas/Page"
          required: false
      responses:
        "200":
          description: Reviews to approve
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    minimum: 0
                  pages:
                    type: integer
                    minimum: 1
                  currentPage:
                    type: integer
                    minimum: 1
                  results:
                    $ref: "#/components/schemas/ListReviews"
                required:
                  - count
                  - pages
                  - currentPage
                  - results
                additionalProperties: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Reviews
  /reviews/{id}:
    parameters:
      - name: id
        description: The unique identifier of the review
        in: path
        schema:
          $ref: "#/components/schemas/Id"
        required: true
    delete:
      operationId: DeleteReview
      description: Delete review
      security:
        - bearerAuth: [ADMINISTRATOR]
      responses:
        "204":
          description: Successfully deleted review
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Reviews
  /reviews/{id}/approve:
    parameters:
      - name: id
        description: The unique identifier of the review
        in: path
        schema:
          $ref: "#/components/schemas/Id"
        required: true
    post:
      operationId: ApproveReview
      description: Approve review
      security:
        - bearerAuth: [ADMINISTRATOR]
      responses:
        "204":
          description: Successfully approved review
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Reviews
...
components:
  responses:
    BadRequest:
      description: Invalid request contents
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Lack of or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: User is not permitted for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: An internal server error has occured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Page:
      description: A listing page number
      type: integer
      minimum: 1
    Id:
      description: The unique identifier
      type: integer
...
    Name:
      type: string
      minLength: 1
      maxLength: 85
    LongName:
      type: string
      minLength: 1
      maxLength: 120
    Description:
      type: string
      minLength: 1
      maxLength: 3000
    Content:
      type: string
      minLength: 1
      maxLength: 1000
    Url:
      type: string
      format: url
      maxLength: 255
...
    UploadType:
      type: string
      enum:
        - FILE
        - IMAGE
...
    Price:
      description: Price of ticket in Wei units
      type: string
      format: ^\d+(.\d{1,2})?$
      examples: ["1000000000000000000.00"]
...
    Date:
      type: string
      format: date
    DateTime:
      type: string
      format: date-time
    Rate:
      type: integer
      minimum: 0
      maximum: 5
...
    CreateArtist:
      type: object
      properties:
        pictureIdUpload:
          $ref: "#/components/schemas/Id"
        name:
          $ref: "#/components/schemas/Name"
        description:
          $ref: "#/components/schemas/Description"
      required:
        - pictureIdUpload
        - name
        - description
      additionalProperties: false
    UpdateArtist:
      $ref: "#/components/schemas/CreateArtist"
...
    GetArtist:
      type: object
      properties:
        idArtist:
          $ref: "#/components/schemas/Id"
        pictureUrl:
          $ref: "#/components/schemas/Url"
        name:
          $ref: "#/components/schemas/Name"
        description:
          $ref: "#/components/schemas/Description"
        events:
          description: List of events in which the artist will perform
          maxItems: 10
          $ref: "#/components/schemas/ListEvents"
        reviews:
          description: List of approved artist's performance reviews
          maxItems: 30
          $ref: "#/components/schemas/ListReviews"
      required:
        - idArtist
        - pictureUrl
        - name
        - description
        - events
        - reviews
      additionalProperties: false
...
    ListEvents:
      type: array
      items:
        type: object
        properties:
          idEvent:
            $ref: "#/components/schemas/Id"
          creatorIdUser:
            $ref: "#/components/schemas/Id"
          creatorUsername:
            $ref: "#/components/schemas/Name"
          idSubcategory:
            $ref: "#/components/schemas/Id"
          subcategoryName:
            $ref: "#/components/schemas/Name"
          idCity:
            $ref: "#/components/schemas/Id"
          cityName:
            $ref: "#/components/schemas/Name"
          name:
            $ref: "#/components/schemas/LongName"
          ticketPrice:
            $ref: "#/components/schemas/Price"
          remainingTicketCount:
            type: integer
            minimum: 0
          location:
            $ref: "#/components/schemas/LongName"
          start:
            $ref: "#/components/schemas/DateTime"
          draft:
            type: boolean
          likes:
            type: integer
            minimum: 0
          transferable:
            type: boolean
          created:
            $ref: "#/components/schemas/DateTime"
          image:
            anyOf:
              - type: "null"
              - $ref: "#/components/schemas/Url"
        required:
          - idEvent
          - creatorIdUser
          - creatorUsername
          - idSubcategory
          - subcategoryName
          - idCity
          - cityName
          - name
          - ticketPrice
          - remainingTicketCount
          - location
          - start
          - likes
          - transferable
          - created
          - image
        additionalProperties: false
...
    CreateUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
        type:
          $ref: "#/components/schemas/UploadType"
      required:
        - file
        - type
      additionalProperties: false
    GetUpload:
      type: object
      properties:
        idUpload:
          $ref: "#/components/schemas/Id"
        url:
          type: string
          format: url
        type:
          $ref: "#/components/schemas/UploadType"
      required:
        - idUpload
        - url
        - type
      additionalProperties: false
...
    ListReviews:
      type: array
      items:
        type: object
        properties:
          idReview:
            $ref: "#/components/schemas/Id"
          reviewerIdUser:
            $ref: "#/components/schemas/Id"
          reviewerUsername:
            $ref: "#/components/schemas/Name"
          reviewedIdArtist:
            $ref: "#/components/schemas/Id"
          reviewedName:
            $ref: "#/components/schemas/Name"
          title:
            $ref: "#/components/schemas/Name"
          eventLocation:
            $ref: "#/components/schemas/LongName"
          eventDate:
            $ref: "#/components/schemas/Date"
          content:
            $ref: "#/components/schemas/Content"
          rate:
            $ref: "#/components/schemas/Rate"
          created:
            $ref: "#/components/schemas/DateTime"
        required:
          - idReview
          - reviewerIdUser
          - reviewerUsername
          - reviewedIdArtist
          - reviewedName
          - title
          - eventLocation
          - eventDate
          - content
          - rate
          - created
        additionalProperties: false
...
    Error:
      type: object
      properties:
        message:
          description: A human readable error message
          type: string
      required:
        - message
      additionalProperties: false
  securitySchemes:
    bearerAuth:
      description: Bearer token using a JWT
      type: http
      scheme: bearer
      bearerFormat: JWT